add core/ and adapters/ for organization
