+++ rust/src/adapters/mod.rs
@@ -1,0 +1,3 @@
+pub mod pyo3_orchestrator;
+
+pub use pyo3_orchestrator::RustLMOrchestrator;
